Refactor FuncXFuture to its own file.
Reduce size of executor, and prevent a circular import for use in
ws_polling_task
