+++ funcx_sdk/funcx/sdk/executor.py
@@ -6,8 +6,8 @@ import queue
 import threading
 import time
 import typing as t
-from concurrent.futures import Future
 
+from funcx.sdk.asynchronous.funcx_future import FuncXFuture
 from funcx.sdk.asynchronous.ws_polling_task import WebSocketPollingTask
 from funcx.sdk.client import FuncXClient
 
@@ -76,19 +76,6 @@ class AtomicController:
 
     def __repr__(self):
         return f"AtomicController value:{self._value}"
-
-
-class FuncXFuture(Future):
-    """Extends concurrent.futures.Future to include an optional task UUID."""
-
-    task_id: t.Optional[str]
-    """The UUID for the task behind this Future. In batch mode, this will
-    not be populated immediately, but will appear later when the task is
-    submitted to the FuncX services."""
-
-    def __init__(self, task_id: t.Optional[str] = None):
-        super().__init__()
-        self.task_id = task_id
 
 
 class FuncXExecutor(concurrent.futures.Executor):
@@ -202,7 +189,7 @@ class FuncXExecutor(concurrent.futures.Executor):
 
         Returns
         -------
-        Future : funcx.sdk.executor.FuncXFuture
+        Future : funcx.sdk.asynchronous.funcx_future.FuncXFuture
             A future object
         """
